customer-service-bot/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM Providers
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# RAG & Vector Store
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Configuration
pyyaml>=6.0.1

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development
black>=24.1.0
ruff>=0.1.14
